v0.3.1023 - /documents leaves client.ts
SCALE-SEAM ⑱. Nine document-control methods (tree, folder, health,
upload/move/delete, download) move to api/documents.ts. One contiguous run.
client.ts 3,353 → 3,304.
