+++ docs/internal/runway-claude-2026-08-19.md
@@ -1,4 +1,4 @@
-# Runway for Claude Code — 2026-08-20, after v0.3.1022
+# Runway for Claude Code — 2026-08-20, after v0.3.1023
 
 **Grade: live handoff.** Written so the next session picks up from measured state rather than chat
 memory. It is **not** the work list — [`docs/roadmap.md`](../roadmap.md) still is. Security
@@ -12,13 +12,13 @@ interface is git: this file, `docs/roadmap.md`, and the stacked PRs. Fetch befor
 
 ## If you are catching main up (Claude Code)
 
-`origin/main` is still **v0.3.986** (`a67c4732`). The Cursor stack is **v0.3.988–1022**, unmerged,
+`origin/main` is still **v0.3.986** (`a67c4732`). The Cursor stack is **v0.3.988–1023**, unmerged,
 draft PRs, each stacked on the previous. Your job is to **land that stack onto main**, not to
 re-implement it.
 
-1. `git fetch origin`. Tip of the stack: `origin/cursor/scale-seam-models-6e15` (**v0.3.1022**).
+1. `git fetch origin`. Tip of the stack: `origin/cursor/scale-seam-documents-6e15` (**v0.3.1023**).
 2. Land **in table order**, one PR at a time. After each merge, rebase the remainder onto the new
-   `main`. **Keep the later version numbers** (do not rebump 1022 down to 987).
+   `main`. **Keep the later version numbers** (do not rebump 1023 down to 987).
 3. Do **not** redo rows in “What already shipped on this stack”.
 4. Do **not** merge #296 (superseded by #297). Do **not** treat #311 as part of this stack
    (cross-origin isolation / nginx — separate).
@@ -60,7 +60,8 @@ re-implement it.
 | 23 | [#316](https://github.com/ibuilder/massing/pull/316) | `cursor/scale-seam-drawing-set-6e15` | **1019** vs #315 |
 | 24 | [#317](https://github.com/ibuilder/massing/pull/317) | `cursor/scale-seam-drawings-6e15` | **1020** vs #316 |
 | 25 | [#318](https://github.com/ibuilder/massing/pull/318) | `cursor/scale-seam-elements-6e15` | **1021** vs #317 |
-| 26 | this | `cursor/scale-seam-models-6e15` | **1022** vs #318 |
+| 26 | [#319](https://github.com/ibuilder/massing/pull/319) | `cursor/scale-seam-models-6e15` | **1022** vs #318 |
+| 27 | this | `cursor/scale-seam-documents-6e15` | **1023** vs #319 |
 
 After each merge: rebase the remainder, **keep the later version numbers**. Do not tag onto a red or
 pending `main`. This agent cannot merge.
@@ -96,6 +97,7 @@ pending `main`. This agent cannot merge.
 | 1020 | SCALE-SEAM ⑮ — `/drawings` mixin; `client.ts` 3,538 → 3,482 |
 | 1021 | SCALE-SEAM ⑯ — `/elements` mixin; `client.ts` 3,482 → 3,412 |
 | 1022 | SCALE-SEAM ⑰ — `/models` mixin; `client.ts` 3,412 → 3,353 |
+| 1023 | SCALE-SEAM ⑱ — `/documents` mixin; `client.ts` 3,353 → 3,304 |
 
 ---
 
@@ -104,7 +106,7 @@ pending `main`. This agent cannot merge.
 - Binding constraint is **adoption / feel**, not missing modules. No second big-ticket.
 - No React, no Reflex. No new npm/PyPI packages without operator OK. Licence MIT/BSD/Apache/ISC/CC0.
 - Do **not** grow `apps/web/src/viewer/app.ts` or `apps/web/src/api/client.ts` without extracting
-  first. Pins: `client.ts` **3,353**, `app.ts` **3,032**, `register.ts` **2,516**.
+  first. Pins: `client.ts` **3,304**, `app.ts` **3,032**, `register.ts` **2,516**.
 - Do **not** turn `AEC_BAKE_SHARE_DIR` on by default. Do **not** add `mapped-diskcache`.
 - Design room brief: **do not add one** (`ROOM_HOME.design` is null).
 - There is **no `pay_app` module** and **no COBie worksheet UI**.
@@ -117,7 +119,7 @@ PERSONA-SHAPE; IDENTITY.
 
 ## Next slices
 
-1. **SCALE-SEAM remainder** — next route-group by size (`/documents`, `/pdf`, `/mep`, `/topics`, …). ⑰ is `/models`.
+1. **SCALE-SEAM remainder** — next route-group by size (`/mep`, `/topics`, `/pdf`, `/ai`, …). ⑱ is `/documents`.
 2. **R24-FIELD-MODE remainder** — replacing the portal home (Lane A). ④ only hides the room tabs.
 3. **R24-REPORTS-BY-MOMENT** remainder — email on a date (SMTP + recipient). Assemble is shipped.
 4. Lane B still open: `R24-TERMS` (user decision), `R22-REPORT-BUILDER` (aggregation/share is backend),
