+++ docs/roadmap.md
@@ -657,7 +657,7 @@ two rows share a path, so two agents in different rows cannot collide.
 | **F · Docs & demo** | `README.md`, `docs/`, `apps/web/src/demo/` | keep the shipped surface honest (below) — no coded items. **`demoData.test.ts` now gates the shell's startup endpoints**; re-run `build_demo_data.py` and that test after adding one |
 | **G · API surface** | `services/api/src/aec_api/routers/`, `main.py` | no standalone items: **every lane routes its own work**, which is why this is a lane rather than a shared file |
 | **H · Registers** | `services/api/modules/*/module.json` | — |
-| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ⑰ |
+| **I · API client** | `apps/web/src/api/` | SCALE-SEAM ⑱ |
 | **J · Build & tooling** | `apps/web/scripts/`, `apps/web/vite.config.ts`, `apps/web/src/style.css`, `apps/web/src/tooling/`, `services/api/test_file_sizes.py`, `services/api/run_tests.py` | — |
 
 **Parked — not available to pick up.** These are decisions or multi-release commitments, listed so
@@ -2753,7 +2753,10 @@ verbs, with a command bar as the escape hatch to everything); and **role-shaped 
   **Design is skipped on purpose:** `apps/web/src/shell/spine.ts` sets Design home to null — the 3D viewer is the room. A portal brief would be a second home.
 ## 🧱 Decomposition & reliability carry-overs (interleave one per few releases)
 
-- ◧ ⭐ **SCALE-SEAM ⑰ — `client.ts` is no longer a god-file, but the split is not finished.** *(⑰ `/models` SHIPPED v0.3.1022; ②–⑯ already shipped)*
+- ◧ ⭐ **SCALE-SEAM ⑱ — `client.ts` is no longer a god-file, but the split is not finished.** *(⑱ `/documents` SHIPPED v0.3.1023; ②–⑰ already shipped)*
+  **⑱ took `/projects/{pid}/documents` out** (9 methods, one contiguous run; `client.ts` 3,353 → 3,304)
+  as `apps/web/src/api/documents.ts`. `api/docqa.ts` stays `/review` + `/doctext`.
+
   **⑰ took `/projects/{pid}/models` out** (9 methods in four regions; `client.ts` 3,412 → 3,353)
   as `apps/web/src/api/models.ts` (`withModels`). `api/model.ts` stays `/model`.
 
@@ -2763,7 +2766,7 @@ verbs, with a command bar as the escape hatch to everything); and **role-shaped 
   **⑮ took `/projects/{pid}/drawings` out** (11 methods in six regions; `client.ts` 3,538 → 3,482)
   as `apps/web/src/api/drawingSheets.ts`. `markupStream` uses `liveStream` on HttpCore.
 
-  **`client.ts` went 4,956 → 3,353 lines** (`wc -l`). Next is the next route-group by size; pick it by
+  **`client.ts` went 4,956 → 3,304 lines** (`wc -l`). Next is the next route-group by size; pick it by
   re-running the classification below, not by reading the section comments.
 
   **This entry read `③+` and named `/model`, `/modules` and `/estimate` as the next groups until
